pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gitsage"
version = "2.2.0"
description = "The Ultimate GitHub Management & Learning Platform - Automate your workflow while mastering GitHub"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "GitSage Contributors"}
]
keywords = ["git", "github", "repository", "management", "wiki", "automation"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: Utilities",
]
requires-python = ">=3.8"
dependencies = [
    "PyYAML>=6.0",
    "rich>=13.0.0",
    "Flask>=3.0.0",
    "Flask-CORS>=4.0.0",
    "GitPython>=3.1.40",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/shadowdevnotreal/gitsage"
Documentation = "https://github.com/shadowdevnotreal/gitsage#readme"
Repository = "https://github.com/shadowdevnotreal/gitsage"
Issues = "https://github.com/shadowdevnotreal/gitsage/issues"
Changelog = "https://github.com/shadowdevnotreal/gitsage/blob/main/CHANGELOG.md"

[project.scripts]
gitsage = "gitsage.cli.launcher:main"
gitsage-web = "gitsage.web.app:main"

[tool.setuptools]

[tool.setuptools.packages.find]
where = ["src"]
include = ["gitsage*"]
exclude = ["tests*"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.package-data]
gitsage = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-v --strict-markers --tb=short"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["src/gitsage"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/env/*",
    "src/gitsage/__version__.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
